+++ src/xcode/ENA/ENA/Source/SceneDelegate.swift
@@ -281,25 +281,8 @@ extension SceneDelegate: ENAExposureManagerObserver {
 		_: ENAExposureManager,
 		didChangeState newState: ExposureManagerState
 	) {
-		func updateHistory() {
-			var history = store.tracingStatusHistory
-			defer {
-				store.tracingStatusHistory = history
-			}
-			let newEntry = TracingStatusEntry(on: newState.isGood, date: Date())
-			guard let lastEntry = history.last else {
-				if newState.isGood {
-					history.append(newEntry)
-				}
-				return
-			}
-			if lastEntry.on != newEntry.on {
-				history.append(newEntry)
-			}
-			store.tracingStatusHistory = history
-		}
-
-		updateHistory()
+		// Add the new state to the history
+		store.tracingStatusHistory = store.tracingStatusHistory.consumingState(newState)
 
 		let message = """
 		New status of EN framework:
